+++ evistream/media/asr/__init__.py
@@ -5,4 +5,3 @@ from evistream.media.asr.mock import MockASR
 from evistream.media.asr.types import ASRAdapter, ASRRequest, ASRResponse, ASRSegment
 
 __all__ = ["ASRAdapter", "ASRRequest", "ASRResponse", "ASRSegment", "FasterWhisperASR", "MockASR"]
-